Fetch metadata for neurons from connectome datasets

Fetch metadata for neurons from connectome datasets

Usage

cf_meta(
  ids,
  bind.rows = TRUE,
  integer64 = FALSE,
  MoreArgs = list(flywire = list(type = c("cell_type", "hemibrain_type")))
)

Arguments

ids

A list of ids named by the relevant datasets (see examples)

bind.rows

Whether to bind data.frames for each dataset together, keeping only the common columns (default TRUE for convenience but note that some columns will be dropped).

integer64

Whether ids should be character vectors (default) or 64 bit ints (more compact but a little fragile as they rely on the bit64 extension package.)

MoreArgs

A named list of arguments to be passed when fetching metadata for a given function. See details.

flywire_type Which metadata column to use from flytable info table

Details

MoreArgs should be list named by the standard dataset names (e.g. as returned by cf_datasets.

See also

Examples

# \donttest{
da2meta=cf_meta(cf_ids(hemibrain='DA2_lPN'))
da2meta
#>           id pre post upstream downstream status    statusLabel    voxels
#> 1 1796817841 396  509      509       3275 Traced Roughly traced 726508881
#> 2 1796818119 511  818      818       4111 Traced Roughly traced 917547898
#> 3 1797505019 345  476      476       2932 Traced Roughly traced 450568345
#> 4 1827516355 391  548      548       3263 Traced Roughly traced 741928779
#> 5  818983130 409  562      562       3400 Traced Roughly traced 612641685
#>   cropped  instance    type cellBodyFiber notes  soma side class group
#> 1   FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE    R  <NA>  <NA>
#> 2   FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE    R  <NA>  <NA>
#> 3   FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE    R  <NA>  <NA>
#> 4   FALSE DA2_lPN_R DA2_lPN         AVM02  <NA>  TRUE    R  <NA>  <NA>
#> 5   FALSE DA2_lPN_R DA2_lPN         AVM02  <NA> FALSE    R  <NA>  <NA>
#>     dataset           key
#> 1 hemibrain hb:1796817841
#> 2 hemibrain hb:1796818119
#> 3 hemibrain hb:1797505019
#> 4 hemibrain hb:1827516355
#> 5 hemibrain  hb:818983130
# / introduces a regular expression
mbonmeta=cf_meta(cf_ids(hemibrain='/MBON.+'))
# }