Specify ids for fly connectome datasets

Specify ids for fly connectome datasets

Usage

cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc"),
  expand = FALSE,
  keys = FALSE,
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL
)

# S3 method for cidlist
c(..., unique = TRUE)

Arguments

query

A query (e.g. cell type name or regular expression)

datasets

Character vector naming datasets to which the query should be applied.

expand

Whether to expand any queries into the matching ids (this will involve one or more calls to corresponding servers). Default FALSE.

keys

Whether to turn the ids into keys hb:12345 right away. Default FALSE but you may find this useful e.g. for combining lists of neurons (see examples).

hemibrain

Pass hemibrain specific query or ids to this argument

flywire

Pass flywire specific query or ids to this argument

malecns

Pass malecns specific query or ids to this argument

manc

Pass manc specific query or ids to this argument

fanc

Pass fanc ids to this argument (at present we do not support metadata queries for fanc)

...

One or more lists generated by cf_ids that should be joined together

unique

Whether to remove duplicated ids when combining cf_id lists with c.cidlist

Value

A list of ids with additional class cidlist

See also

Examples

# \donttest{
cf_ids("DA2_lPN", datasets='brain')
#> flywire [1 ids]: DA2_lPN
#> hemibrain [1 ids]: DA2_lPN
#> malecns [1 ids]: DA2_lPN
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')
#> flywire [1 ids]: /MBON.+
#> hemibrain [1 ids]: /MBON.+
#> malecns [1 ids]: /MBON.+

# expand query into actual ids
cf_ids("/type:MBON.+", datasets='brain', expand=TRUE)
#> Warning: No valid ids in dataset:flywire
#> Error in clio_auth() : 
#>   Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
#> Warning: Trouble choosing default malecns dataset.
#> Try running dr_malecns() and then ask on #code or file an issue at
#> https://github.com/flyconnectome/malecns/issues
#> Warning: No valid ids in dataset:malecns
#> flywire [0 ids]: 
#> hemibrain [71 ids]: 612371421 673509195 424789697 5813022341 487925037 612738462 487925063 613079053 1386512867 799586652 ...
#> malecns [0 ids]: 

# return keys directly
cf_ids("/type:MBON.+", keys=TRUE)
#> Warning: No valid ids in dataset:flywire
#> Warning: No valid ids in dataset:malecns
#> hb:612371421
#> hb:673509195
#> hb:424789697
#> hb:5813022341
#> hb:487925037
#> hb:612738462
#> hb:487925063
#> hb:613079053
#> hb:1386512867
#> hb:799586652
#> hb:422725634
#> hb:5813061512
#> hb:1234386037
#> hb:611477605
#> hb:733036127
#> hb:920470959
#> hb:1016835041
#> hb:1048215779
#> hb:1078693835
#> hb:613719036
#> hb:672352543
#> hb:768555687
#> hb:424767514
#> hb:517518166
#> hb:704466265
#> hb:861665641
#> hb:1139667240
#> hb:300972942
#> hb:394225044
#> hb:5813067721
#> hb:5813068729
#> hb:642664141
#> hb:673702721
#> hb:457175171
#> hb:579916831
#> hb:673366098
#> hb:5812981264
#> hb:5812981543
#> hb:457196444
#> hb:5813020828
#> hb:423774471
#> hb:517854468
#> hb:792368888
#> hb:5813022896
#> hb:985813153
#> hb:5813042659
#> hb:985131499
#> hb:423382015
#> hb:487143497
#> hb:550081497
#> hb:612742248
#> hb:612422902
#> hb:912951014
#> hb:895441451
#> hb:5813061538
#> hb:5813117385
#> hb:457196643
#> hb:5813058048
#> hb:5813052909
#> hb:5813022234
#> hb:5813040205
#> hb:706948318
#> hb:894020730
#> hb:642732835
#> hb:643847447
#> hb:518670779
#> hb:547552266
#> hb:1173814525
#> hb:5813033000
#> hb:518930199
#> hb:611089775
# one way of combining separate lists of neurons
hbids=c(264083994, 5813022274)
c(cf_ids("/type:MBON1.+", keys=TRUE), cf_ids(hemibrain = hbids, keys = TRUE))
#> Warning: No valid ids in dataset:flywire
#> Warning: No valid ids in dataset:malecns
#>  [1] "hb:613719036"  "hb:672352543"  "hb:768555687"  "hb:424767514" 
#>  [5] "hb:517518166"  "hb:704466265"  "hb:861665641"  "hb:1139667240"
#>  [9] "hb:300972942"  "hb:394225044"  "hb:5813067721" "hb:5813068729"
#> [13] "hb:642664141"  "hb:673702721"  "hb:457175171"  "hb:579916831" 
#> [17] "hb:673366098"  "hb:5812981264" "hb:5812981543" "hb:457196444" 
#> [21] "hb:5813020828" "hb:423774471"  "hb:517854468"  "hb:264083994" 
#> [25] "hb:5813022274"

# now equivalent to
keys(c(cf_ids("/type:MBON1.+"), cf_ids(hemibrain = hbids)))
#> Warning: No valid ids in dataset:flywire
#> Warning: No valid ids in dataset:malecns
#> hb:613719036
#> hb:672352543
#> hb:768555687
#> hb:424767514
#> hb:517518166
#> hb:704466265
#> hb:861665641
#> hb:1139667240
#> hb:300972942
#> hb:394225044
#> hb:5813067721
#> hb:5813068729
#> hb:642664141
#> hb:673702721
#> hb:457175171
#> hb:579916831
#> hb:673366098
#> hb:5812981264
#> hb:5812981543
#> hb:457196444
#> hb:5813020828
#> hb:423774471
#> hb:517854468
#> hb:264083994
#> hb:5813022274
# }
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [2 ids]: 2 1
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2), unique=FALSE)
#> flywire [3 ids]: 1 1 2
#> hemibrain [3 ids]: 2 1 2
# \donttest{
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = '/SMP55[0-9]', flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [19 ids]: 2 452689494 579700707 328274638 392757094 605166418 298262663 361114148 669670654 613394707 ...
# }