Specify ids for fly connectome datasets

Specify ids for fly connectome datasets

Usage

cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc"),
  expand = FALSE,
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL
)

Arguments

query

A query (e.g. cell type name or regular expression)

datasets

Character vector naming datasets to which the query should be applied.

expand

Whether to expand any queries into the matching ids (this will involve one or more calls to corresponding servers). Default FALSE.

hemibrain

Pass hemibrain specific query or ids to this argument

flywire

Pass flywire specific query or ids to this argument

malecns

Pass malecns specific query or ids to this argument

manc

Pass manc specific query or ids to this argument

fanc

Pass fanc ids to this argument (at present we do not support metadata queries for fanc)

Value

A list of ids

See also

Examples

# \donttest{
cf_ids("DA2_lPN", datasets='brain')
#> $hemibrain
#> [1] "DA2_lPN"
#> 
#> $flywire
#> [1] "DA2_lPN"
#> 
#> $malecns
#> [1] "DA2_lPN"
#> 
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')
#> $hemibrain
#> [1] "/MBON.+"
#> 
#> $flywire
#> [1] "/MBON.+"
#> 
#> $malecns
#> [1] "/MBON.+"
#> 

# expand query into actual ids
cf_ids("/type:MBON.+", datasets='brain', expand=TRUE)
#> Warning: No valid ids in dataset:flywire
#> Error in loadNamespace(x): there is no package called ‘malecns’
# }