Specify ids for fly connectome datasets

Specify ids for fly connectome datasets

Usage

cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc",
    "opticlobe"),
  expand = FALSE,
  keys = FALSE,
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL,
  opticlobe = NULL
)

# S3 method for cidlist
c(..., unique = TRUE)

Arguments

query

A query (e.g. cell type name or regular expression)

datasets

Character vector naming datasets to which the query should be applied.

expand

Whether to expand any queries into the matching ids (this will involve one or more calls to corresponding servers). Default FALSE.

keys

Whether to turn the ids into keys hb:12345 right away. Default FALSE but you may find this useful e.g. for combining lists of neurons (see examples).

hemibrain

Pass hemibrain specific query or ids to this argument

flywire

Pass flywire specific query or ids to this argument

malecns

Pass malecns specific query or ids to this argument

manc

Pass manc specific query or ids to this argument

fanc

Pass fanc ids to this argument (at present we do not support metadata queries for fanc)

opticlobe

Pass opticlobe specific query or ids to this argument

...

One or more lists generated by cf_ids that should be joined together

unique

Whether to remove duplicated ids when combining cf_id lists with c.cidlist

Value

A list of ids with additional class cidlist

Details

all neuprint datasets (hemibrain, malevnc, opticlobe, malecns) use the same query syntax although some fields may be dataset specific (see examples).

See also

Examples

# \donttest{
cf_ids("DA2_lPN", datasets='brain')
#> flywire [1 ids]: DA2_lPN
#> hemibrain [1 ids]: DA2_lPN
#> malecns [1 ids]: DA2_lPN
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')
#> flywire [1 ids]: /MBON.+
#> hemibrain [1 ids]: /MBON.+
#> malecns [1 ids]: /MBON.+

# expand query into actual ids
cf_ids("/type:MBON.+", datasets='brain', expand=TRUE)
#> Error in clio_auth() : 
#>   Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
#> Warning: Trouble choosing default malecns dataset.
#> Try running dr_malecns() and then ask on #code or file an issue at
#> https://github.com/flyconnectome/malecns/issues
#> Warning: No valid ids in dataset:malecns
#> flywire [96 ids]: 720575940624185095 720575940624117245 720575940638526278 720575940629585602 720575940637934308 720575940632118343 720575940609959637 720575940638163428 720575940630496374 720575940642142861 ...
#> hemibrain [71 ids]: 612371421 673509195 424789697 5813022341 487925037 612738462 487925063 613079053 1386512867 799586652 ...
#> malecns [0 ids]: 

# return keys directly
cf_ids("/type:MBON.+", keys=TRUE)
#> Warning: No valid ids in dataset:malecns
#>   [1] "fw:720575940624185095" "fw:720575940624117245" "fw:720575940638526278"
#>   [4] "fw:720575940629585602" "fw:720575940637934308" "fw:720575940632118343"
#>   [7] "fw:720575940609959637" "fw:720575940638163428" "fw:720575940630496374"
#>  [10] "fw:720575940642142861" "fw:720575940611344078" "fw:720575940634822751"
#>  [13] "fw:720575940617552340" "fw:720575940624280328" "fw:720575940617567206"
#>  [16] "fw:720575940643700640" "fw:720575940615221811" "fw:720575940614026193"
#>  [19] "fw:720575940633209647" "fw:720575940623001480" "fw:720575940608304220"
#>  [22] "fw:720575940614595218" "fw:720575940636992368" "fw:720575940621828443"
#>  [25] "fw:720575940637902938" "fw:720575940638774606" "fw:720575940639697827"
#>  [28] "fw:720575940626833021" "fw:720575940618008859" "fw:720575940635841143"
#>  [31] "fw:720575940618249797" "fw:720575940629856515" "fw:720575940623905719"
#>  [34] "fw:720575940629529722" "fw:720575940624659943" "fw:720575940624590316"
#>  [37] "fw:720575940645304430" "fw:720575940612196850" "fw:720575940630767959"
#>  [40] "fw:720575940635082347" "fw:720575940628757547" "fw:720575940627144069"
#>  [43] "fw:720575940610647416" "fw:720575940622093436" "fw:720575940624696810"
#>  [46] "fw:720575940629422086" "fw:720575940647328900" "fw:720575940650386553"
#>  [49] "fw:720575940639556467" "fw:720575940612455338" "fw:720575940617749538"
#>  [52] "fw:720575940617760257" "fw:720575940619810389" "fw:720575940634482783"
#>  [55] "fw:720575940644615716" "fw:720575940626109315" "fw:720575940623377802"
#>  [58] "fw:720575940626744921" "fw:720575940623464316" "fw:720575940622979277"
#>  [61] "fw:720575940614892182" "fw:720575940620715224" "fw:720575940643863496"
#>  [64] "fw:720575940621777391" "fw:720575940606953858" "fw:720575940630075703"
#>  [67] "fw:720575940652390134" "fw:720575940632535756" "fw:720575940643696288"
#>  [70] "fw:720575940623201833" "fw:720575940635063135" "fw:720575940616463477"
#>  [73] "fw:720575940629981440" "fw:720575940623182847" "fw:720575940626315010"
#>  [76] "fw:720575940632943277" "fw:720575940624539284" "fw:720575940620464321"
#>  [79] "fw:720575940628783363" "fw:720575940643309197" "fw:720575940614392999"
#>  [82] "fw:720575940617302365" "fw:720575940622997453" "fw:720575940624694503"
#>  [85] "fw:720575940628734376" "fw:720575940610964946" "fw:720575940607155890"
#>  [88] "fw:720575940623381956" "fw:720575940630864847" "fw:720575940638028607"
#>  [91] "fw:720575940621164720" "fw:720575940623841463" "fw:720575940631177803"
#>  [94] "fw:720575940624489964" "fw:720575940616398303" "fw:720575940623750325"
#>  [97] "hb:612371421"          "hb:673509195"          "hb:424789697"         
#> [100] "hb:5813022341"         "hb:487925037"          "hb:612738462"         
#> [103] "hb:487925063"          "hb:613079053"          "hb:1386512867"        
#> [106] "hb:799586652"          "hb:422725634"          "hb:5813061512"        
#> [109] "hb:1234386037"         "hb:611477605"          "hb:733036127"         
#> [112] "hb:920470959"          "hb:1016835041"         "hb:1048215779"        
#> [115] "hb:1078693835"         "hb:613719036"          "hb:672352543"         
#> [118] "hb:768555687"          "hb:424767514"          "hb:517518166"         
#> [121] "hb:704466265"          "hb:861665641"          "hb:1139667240"        
#> [124] "hb:300972942"          "hb:394225044"          "hb:5813067721"        
#> [127] "hb:5813068729"         "hb:642664141"          "hb:673702721"         
#> [130] "hb:457175171"          "hb:579916831"          "hb:673366098"         
#> [133] "hb:5812981264"         "hb:5812981543"         "hb:457196444"         
#> [136] "hb:5813020828"         "hb:423774471"          "hb:517854468"         
#> [139] "hb:792368888"          "hb:5813022896"         "hb:985813153"         
#> [142] "hb:5813042659"         "hb:985131499"          "hb:423382015"         
#> [145] "hb:487143497"          "hb:550081497"          "hb:612742248"         
#> [148] "hb:612422902"          "hb:912951014"          "hb:895441451"         
#> [151] "hb:5813061538"         "hb:5813117385"         "hb:457196643"         
#> [154] "hb:5813058048"         "hb:5813052909"         "hb:5813022234"        
#> [157] "hb:5813040205"         "hb:706948318"          "hb:894020730"         
#> [160] "hb:642732835"          "hb:643847447"          "hb:518670779"         
#> [163] "hb:547552266"          "hb:1173814525"         "hb:5813033000"        
#> [166] "hb:518930199"          "hb:611089775"         
# one way of combining separate lists of neurons
hbids=c(264083994, 5813022274)
c(cf_ids("/type:MBON1.+", keys=TRUE), cf_ids(hemibrain = hbids, keys = TRUE))
#> Warning: No valid ids in dataset:malecns
#>  [1] "fw:720575940629585602" "fw:720575940632118343" "fw:720575940642142861"
#>  [4] "fw:720575940634822751" "fw:720575940615221811" "fw:720575940614026193"
#>  [7] "fw:720575940633209647" "fw:720575940623001480" "fw:720575940608304220"
#> [10] "fw:720575940621828443" "fw:720575940638774606" "fw:720575940639697827"
#> [13] "fw:720575940626833021" "fw:720575940635841143" "fw:720575940623905719"
#> [16] "fw:720575940629529722" "fw:720575940645304430" "fw:720575940612196850"
#> [19] "fw:720575940630767959" "fw:720575940635082347" "fw:720575940628757547"
#> [22] "fw:720575940610647416" "fw:720575940622093436" "fw:720575940624696810"
#> [25] "fw:720575940647328900" "fw:720575940650386553" "fw:720575940639556467"
#> [28] "fw:720575940612455338" "fw:720575940617749538" "fw:720575940617760257"
#> [31] "fw:720575940634482783" "fw:720575940626109315" "fw:720575940623377802"
#> [34] "fw:720575940626744921" "fw:720575940630075703" "fw:720575940632535756"
#> [37] "fw:720575940643696288" "fw:720575940623201833" "fw:720575940626315010"
#> [40] "fw:720575940624539284" "fw:720575940622997453" "fw:720575940623841463"
#> [43] "hb:613719036"          "hb:672352543"          "hb:768555687"         
#> [46] "hb:424767514"          "hb:517518166"          "hb:704466265"         
#> [49] "hb:861665641"          "hb:1139667240"         "hb:300972942"         
#> [52] "hb:394225044"          "hb:5813067721"         "hb:5813068729"        
#> [55] "hb:642664141"          "hb:673702721"          "hb:457175171"         
#> [58] "hb:579916831"          "hb:673366098"          "hb:5812981264"        
#> [61] "hb:5812981543"         "hb:457196444"          "hb:5813020828"        
#> [64] "hb:423774471"          "hb:517854468"          "hb:264083994"         
#> [67] "hb:5813022274"        

# now equivalent to
keys(c(cf_ids("/type:MBON1.+"), cf_ids(hemibrain = hbids)))
#> Warning: No valid ids in dataset:malecns
#>  [1] "fw:720575940629585602" "fw:720575940632118343" "fw:720575940642142861"
#>  [4] "fw:720575940634822751" "fw:720575940615221811" "fw:720575940614026193"
#>  [7] "fw:720575940633209647" "fw:720575940623001480" "fw:720575940608304220"
#> [10] "fw:720575940621828443" "fw:720575940638774606" "fw:720575940639697827"
#> [13] "fw:720575940626833021" "fw:720575940635841143" "fw:720575940623905719"
#> [16] "fw:720575940629529722" "fw:720575940645304430" "fw:720575940612196850"
#> [19] "fw:720575940630767959" "fw:720575940635082347" "fw:720575940628757547"
#> [22] "fw:720575940610647416" "fw:720575940622093436" "fw:720575940624696810"
#> [25] "fw:720575940647328900" "fw:720575940650386553" "fw:720575940639556467"
#> [28] "fw:720575940612455338" "fw:720575940617749538" "fw:720575940617760257"
#> [31] "fw:720575940634482783" "fw:720575940626109315" "fw:720575940623377802"
#> [34] "fw:720575940626744921" "fw:720575940630075703" "fw:720575940632535756"
#> [37] "fw:720575940643696288" "fw:720575940623201833" "fw:720575940626315010"
#> [40] "fw:720575940624539284" "fw:720575940622997453" "fw:720575940623841463"
#> [43] "hb:613719036"          "hb:672352543"          "hb:768555687"         
#> [46] "hb:424767514"          "hb:517518166"          "hb:704466265"         
#> [49] "hb:861665641"          "hb:1139667240"         "hb:300972942"         
#> [52] "hb:394225044"          "hb:5813067721"         "hb:5813068729"        
#> [55] "hb:642664141"          "hb:673702721"          "hb:457175171"         
#> [58] "hb:579916831"          "hb:673366098"          "hb:5812981264"        
#> [61] "hb:5812981543"         "hb:457196444"          "hb:5813020828"        
#> [64] "hb:423774471"          "hb:517854468"          "hb:264083994"         
#> [67] "hb:5813022274"        
# }
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [2 ids]: 2 1
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2), unique=FALSE)
#> flywire [3 ids]: 1 1 2
#> hemibrain [3 ids]: 2 1 2
# \donttest{
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = '/SMP55[0-9]', flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [19 ids]: 2 452689494 579700707 328274638 392757094 605166418 298262663 361114148 669670654 613394707 ...
# }