Specify ids for fly connectome datasets

Specify ids for fly connectome datasets

Usage

cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc",
    "opticlobe", "banc"),
  expand = FALSE,
  keys = FALSE,
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL,
  opticlobe = NULL,
  banc = NULL
)

# S3 method for class 'cidlist'
c(..., unique = TRUE)

Arguments

query

A query (e.g. cell type name or regular expression)

datasets

Character vector naming datasets to which the query should be applied.

expand

Whether to expand any queries into the matching ids (this will involve one or more calls to corresponding servers). Default FALSE.

keys

Whether to turn the ids into keys hb:12345 right away. Default FALSE but you may find this useful e.g. for combining lists of neurons (see examples).

hemibrain

Pass hemibrain specific query or ids to this argument

flywire

Pass flywire specific query or ids to this argument

malecns

Pass malecns specific query or ids to this argument

manc

Pass manc specific query or ids to this argument

fanc

Pass fanc ids to this argument (at present we do not support metadata queries for fanc)

opticlobe

Pass opticlobe specific query or ids to this argument

banc

Pass banc ids to this argument (we only support basic metadata queries for banc)

...

One or more lists generated by cf_ids that should be joined together

unique

Whether to remove duplicated ids when combining cf_id lists with c.cidlist

Value

A list of ids with additional class cidlist

Details

You will often want to perform a query, most commonly for a cell type or cell class, rather than specific numeric ids. The most flexible way to do this is to use a regular expression (regex) query, specified with an initial "/".

All neuprint datasets (hemibrain, malevnc, opticlobe, malecns) use the same query syntax although some fields may be dataset specific (see examples). The regex syntax for CAVE datasets (flywire, fanc, banc) should be the same although you may find some wrinkles because the underlying data stores are different. Note that we do not yet translate all the different fields across datasets for queries, although this is a goal. For example the neuprint/fanc/banc class field is equivalent to flywire super_class. Similarly the values are not guaranteed to be the same. Where flywire uses super_class=="descending") the manc uses class=="descending neuron").

Therefore to find all DNs in these two datasets you will need to do: cf_ids(manc='/class:descending.*', flywire='/super_class:descending.*')

Feel free to make an issue if you find something that doesn't feel right or can suggest an improvement.

See also

Examples

# \donttest{
cf_ids("DA2_lPN", datasets='brain')
#> banc [1 ids]: DA2_lPN
#> flywire [1 ids]: DA2_lPN
#> hemibrain [1 ids]: DA2_lPN
#> malecns [1 ids]: DA2_lPN
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')
#> banc [1 ids]: /MBON.+
#> flywire [1 ids]: /MBON.+
#> hemibrain [1 ids]: /MBON.+
#> malecns [1 ids]: /MBON.+

# expand query into actual ids
cf_ids("/type:MBON.+", datasets='brain', expand=TRUE)
#> Error in clio_auth() : 
#>   Clio/Google auth failure. Do you have access rights to VNC clio?
#> Try specifying the email linked to clio in a call to `clio_auth` or setting `options(malevnc.clio_email)`!
#> Warning: Trouble choosing default malecns dataset.
#> Try running dr_malecns() and then ask on #code or file an issue at
#> https://github.com/flyconnectome/malecns/issues
#> Warning: Unable to process query for dataset:malecns
#> Warning: WARNING:root:Deprecation: this method is to facilitate beta testing of this feature,             it will likely get removed in future versions. 
#> flywire_rootid_cached: Looking up 13 missing keys
#> banc [65 ids]: 720575941400612387 720575941407672494 720575941413591697 720575941426934100 720575941428592559 720575941430336137 720575941432355086 720575941458631494 720575941461639312 720575941467001622 ...
#> flywire [96 ids]: 720575940624185095 720575940624117245 720575940638526278 720575940629585602 720575940637934308 720575940632118343 720575940609959637 720575940638163428 720575940630496374 720575940642142861 ...
#> hemibrain [71 ids]: 612371421 673509195 424789697 5813022341 487925037 612738462 487925063 613079053 1386512867 799586652 ...
#> malecns [0 ids]: 

# return keys directly
cf_ids("/type:MBON.+", keys=TRUE)
#> Warning: Unable to process query for dataset:malecns
#>   [1] "bc:720575941400612387" "bc:720575941407672494" "bc:720575941413591697"
#>   [4] "bc:720575941426934100" "bc:720575941428592559" "bc:720575941430336137"
#>   [7] "bc:720575941432355086" "bc:720575941458631494" "bc:720575941461639312"
#>  [10] "bc:720575941467001622" "bc:720575941475882545" "bc:720575941477553002"
#>  [13] "bc:720575941479618979" "bc:720575941485799430" "bc:720575941487270608"
#>  [16] "bc:720575941489258441" "bc:720575941489773621" "bc:720575941490068169"
#>  [19] "bc:720575941492491440" "bc:720575941493981822" "bc:720575941494355070"
#>  [22] "bc:720575941498066249" "bc:720575941498202593" "bc:720575941506241815"
#>  [25] "bc:720575941510013394" "bc:720575941514838521" "bc:720575941515157850"
#>  [28] "bc:720575941517355180" "bc:720575941518103724" "bc:720575941518849134"
#>  [31] "bc:720575941520068787" "bc:720575941522167031" "bc:720575941529029048"
#>  [34] "bc:720575941531794149" "bc:720575941539422669" "bc:720575941541784156"
#>  [37] "bc:720575941544051973" "bc:720575941545226676" "bc:720575941546111880"
#>  [40] "bc:720575941549399156" "bc:720575941550979327" "bc:720575941553055515"
#>  [43] "bc:720575941553207323" "bc:720575941553361691" "bc:720575941559025134"
#>  [46] "bc:720575941559770739" "bc:720575941562860690" "bc:720575941564924055"
#>  [49] "bc:720575941567268090" "bc:720575941573578235" "bc:720575941575148982"
#>  [52] "bc:720575941575586664" "bc:720575941578741073" "bc:720575941592415339"
#>  [55] "bc:720575941597599644" "bc:720575941599755049" "bc:720575941607000429"
#>  [58] "bc:720575941611287014" "bc:720575941628990758" "bc:720575941631918496"
#>  [61] "bc:720575941643022162" "bc:720575941662583676" "bc:720575941687317231"
#>  [64] "bc:720575941690272664" "bc:720575941722399802" "fw:720575940624185095"
#>  [67] "fw:720575940624117245" "fw:720575940638526278" "fw:720575940629585602"
#>  [70] "fw:720575940637934308" "fw:720575940632118343" "fw:720575940609959637"
#>  [73] "fw:720575940638163428" "fw:720575940630496374" "fw:720575940642142861"
#>  [76] "fw:720575940611344078" "fw:720575940634822751" "fw:720575940617552340"
#>  [79] "fw:720575940624280328" "fw:720575940617567206" "fw:720575940643700640"
#>  [82] "fw:720575940615221811" "fw:720575940614026193" "fw:720575940633209647"
#>  [85] "fw:720575940623001480" "fw:720575940608304220" "fw:720575940614595218"
#>  [88] "fw:720575940636992368" "fw:720575940621828443" "fw:720575940637902938"
#>  [91] "fw:720575940638774606" "fw:720575940639697827" "fw:720575940626833021"
#>  [94] "fw:720575940618008859" "fw:720575940635841143" "fw:720575940618249797"
#>  [97] "fw:720575940629856515" "fw:720575940623905719" "fw:720575940629529722"
#> [100] "fw:720575940624659943" "fw:720575940624590316" "fw:720575940645304430"
#> [103] "fw:720575940612196850" "fw:720575940630767959" "fw:720575940635082347"
#> [106] "fw:720575940628757547" "fw:720575940627144069" "fw:720575940610647416"
#> [109] "fw:720575940622093436" "fw:720575940624696810" "fw:720575940629422086"
#> [112] "fw:720575940647328900" "fw:720575940650386553" "fw:720575940639556467"
#> [115] "fw:720575940612455338" "fw:720575940617749538" "fw:720575940617760257"
#> [118] "fw:720575940619810389" "fw:720575940634482783" "fw:720575940644615716"
#> [121] "fw:720575940626109315" "fw:720575940623377802" "fw:720575940626744921"
#> [124] "fw:720575940623464316" "fw:720575940622979277" "fw:720575940614892182"
#> [127] "fw:720575940620715224" "fw:720575940643863496" "fw:720575940621777391"
#> [130] "fw:720575940606953858" "fw:720575940630075703" "fw:720575940652390134"
#> [133] "fw:720575940632535756" "fw:720575940643696288" "fw:720575940623201833"
#> [136] "fw:720575940635063135" "fw:720575940616463477" "fw:720575940629981440"
#> [139] "fw:720575940623182847" "fw:720575940626315010" "fw:720575940632943277"
#> [142] "fw:720575940624539284" "fw:720575940620464321" "fw:720575940628783363"
#> [145] "fw:720575940643309197" "fw:720575940614392999" "fw:720575940617302365"
#> [148] "fw:720575940622997453" "fw:720575940624694503" "fw:720575940628734376"
#> [151] "fw:720575940610964946" "fw:720575940607155890" "fw:720575940623381956"
#> [154] "fw:720575940630864847" "fw:720575940638028607" "fw:720575940621164720"
#> [157] "fw:720575940623841463" "fw:720575940631177803" "fw:720575940624489964"
#> [160] "fw:720575940616398303" "fw:720575940623750325" "hb:612371421"         
#> [163] "hb:673509195"          "hb:424789697"          "hb:5813022341"        
#> [166] "hb:487925037"          "hb:612738462"          "hb:487925063"         
#> [169] "hb:613079053"          "hb:1386512867"         "hb:799586652"         
#> [172] "hb:422725634"          "hb:5813061512"         "hb:1234386037"        
#> [175] "hb:611477605"          "hb:733036127"          "hb:920470959"         
#> [178] "hb:1016835041"         "hb:1048215779"         "hb:1078693835"        
#> [181] "hb:613719036"          "hb:672352543"          "hb:768555687"         
#> [184] "hb:424767514"          "hb:517518166"          "hb:704466265"         
#> [187] "hb:861665641"          "hb:1139667240"         "hb:300972942"         
#> [190] "hb:394225044"          "hb:5813067721"         "hb:5813068729"        
#> [193] "hb:642664141"          "hb:673702721"          "hb:457175171"         
#> [196] "hb:579916831"          "hb:673366098"          "hb:5812981264"        
#> [199] "hb:5812981543"         "hb:457196444"          "hb:5813020828"        
#> [202] "hb:423774471"          "hb:517854468"          "hb:792368888"         
#> [205] "hb:5813022896"         "hb:985813153"          "hb:5813042659"        
#> [208] "hb:985131499"          "hb:423382015"          "hb:487143497"         
#> [211] "hb:550081497"          "hb:612742248"          "hb:612422902"         
#> [214] "hb:912951014"          "hb:895441451"          "hb:5813061538"        
#> [217] "hb:5813117385"         "hb:457196643"          "hb:5813058048"        
#> [220] "hb:5813052909"         "hb:5813022234"         "hb:5813040205"        
#> [223] "hb:706948318"          "hb:894020730"          "hb:642732835"         
#> [226] "hb:643847447"          "hb:518670779"          "hb:547552266"         
#> [229] "hb:1173814525"         "hb:5813033000"         "hb:518930199"         
#> [232] "hb:611089775"         
# one way of combining separate lists of neurons
hbids=c(264083994, 5813022274)
c(cf_ids("/type:MBON1.+", keys=TRUE), cf_ids(hemibrain = hbids, keys = TRUE))
#> Warning: Unable to process query for dataset:malecns
#>   [1] "bc:720575941400612387" "bc:720575941407672494" "bc:720575941413591697"
#>   [4] "bc:720575941426934100" "bc:720575941428592559" "bc:720575941430336137"
#>   [7] "bc:720575941461639312" "bc:720575941475882545" "bc:720575941479618979"
#>  [10] "bc:720575941489773621" "bc:720575941493981822" "bc:720575941494355070"
#>  [13] "bc:720575941498066249" "bc:720575941498202593" "bc:720575941506241815"
#>  [16] "bc:720575941510013394" "bc:720575941514838521" "bc:720575941517355180"
#>  [19] "bc:720575941518103724" "bc:720575941520068787" "bc:720575941522167031"
#>  [22] "bc:720575941539422669" "bc:720575941545226676" "bc:720575941546111880"
#>  [25] "bc:720575941549399156" "bc:720575941550979327" "bc:720575941553207323"
#>  [28] "bc:720575941553361691" "bc:720575941559025134" "bc:720575941578741073"
#>  [31] "bc:720575941631918496" "bc:720575941687317231" "bc:720575941690272664"
#>  [34] "bc:720575941722399802" "fw:720575940629585602" "fw:720575940632118343"
#>  [37] "fw:720575940642142861" "fw:720575940634822751" "fw:720575940615221811"
#>  [40] "fw:720575940614026193" "fw:720575940633209647" "fw:720575940623001480"
#>  [43] "fw:720575940608304220" "fw:720575940621828443" "fw:720575940638774606"
#>  [46] "fw:720575940639697827" "fw:720575940626833021" "fw:720575940635841143"
#>  [49] "fw:720575940623905719" "fw:720575940629529722" "fw:720575940645304430"
#>  [52] "fw:720575940612196850" "fw:720575940630767959" "fw:720575940635082347"
#>  [55] "fw:720575940628757547" "fw:720575940610647416" "fw:720575940622093436"
#>  [58] "fw:720575940624696810" "fw:720575940647328900" "fw:720575940650386553"
#>  [61] "fw:720575940639556467" "fw:720575940612455338" "fw:720575940617749538"
#>  [64] "fw:720575940617760257" "fw:720575940634482783" "fw:720575940626109315"
#>  [67] "fw:720575940623377802" "fw:720575940626744921" "fw:720575940630075703"
#>  [70] "fw:720575940632535756" "fw:720575940643696288" "fw:720575940623201833"
#>  [73] "fw:720575940626315010" "fw:720575940624539284" "fw:720575940622997453"
#>  [76] "fw:720575940623841463" "hb:613719036"          "hb:672352543"         
#>  [79] "hb:768555687"          "hb:424767514"          "hb:517518166"         
#>  [82] "hb:704466265"          "hb:861665641"          "hb:1139667240"        
#>  [85] "hb:300972942"          "hb:394225044"          "hb:5813067721"        
#>  [88] "hb:5813068729"         "hb:642664141"          "hb:673702721"         
#>  [91] "hb:457175171"          "hb:579916831"          "hb:673366098"         
#>  [94] "hb:5812981264"         "hb:5812981543"         "hb:457196444"         
#>  [97] "hb:5813020828"         "hb:423774471"          "hb:517854468"         
#> [100] "hb:264083994"          "hb:5813022274"        

# now equivalent to
keys(c(cf_ids("/type:MBON1.+"), cf_ids(hemibrain = hbids)))
#> Warning: Unable to process query for dataset:malecns
#>   [1] "bc:720575941400612387" "bc:720575941407672494" "bc:720575941413591697"
#>   [4] "bc:720575941426934100" "bc:720575941428592559" "bc:720575941430336137"
#>   [7] "bc:720575941461639312" "bc:720575941475882545" "bc:720575941479618979"
#>  [10] "bc:720575941489773621" "bc:720575941493981822" "bc:720575941494355070"
#>  [13] "bc:720575941498066249" "bc:720575941498202593" "bc:720575941506241815"
#>  [16] "bc:720575941510013394" "bc:720575941514838521" "bc:720575941517355180"
#>  [19] "bc:720575941518103724" "bc:720575941520068787" "bc:720575941522167031"
#>  [22] "bc:720575941539422669" "bc:720575941545226676" "bc:720575941546111880"
#>  [25] "bc:720575941549399156" "bc:720575941550979327" "bc:720575941553207323"
#>  [28] "bc:720575941553361691" "bc:720575941559025134" "bc:720575941578741073"
#>  [31] "bc:720575941631918496" "bc:720575941687317231" "bc:720575941690272664"
#>  [34] "bc:720575941722399802" "fw:720575940629585602" "fw:720575940632118343"
#>  [37] "fw:720575940642142861" "fw:720575940634822751" "fw:720575940615221811"
#>  [40] "fw:720575940614026193" "fw:720575940633209647" "fw:720575940623001480"
#>  [43] "fw:720575940608304220" "fw:720575940621828443" "fw:720575940638774606"
#>  [46] "fw:720575940639697827" "fw:720575940626833021" "fw:720575940635841143"
#>  [49] "fw:720575940623905719" "fw:720575940629529722" "fw:720575940645304430"
#>  [52] "fw:720575940612196850" "fw:720575940630767959" "fw:720575940635082347"
#>  [55] "fw:720575940628757547" "fw:720575940610647416" "fw:720575940622093436"
#>  [58] "fw:720575940624696810" "fw:720575940647328900" "fw:720575940650386553"
#>  [61] "fw:720575940639556467" "fw:720575940612455338" "fw:720575940617749538"
#>  [64] "fw:720575940617760257" "fw:720575940634482783" "fw:720575940626109315"
#>  [67] "fw:720575940623377802" "fw:720575940626744921" "fw:720575940630075703"
#>  [70] "fw:720575940632535756" "fw:720575940643696288" "fw:720575940623201833"
#>  [73] "fw:720575940626315010" "fw:720575940624539284" "fw:720575940622997453"
#>  [76] "fw:720575940623841463" "hb:613719036"          "hb:672352543"         
#>  [79] "hb:768555687"          "hb:424767514"          "hb:517518166"         
#>  [82] "hb:704466265"          "hb:861665641"          "hb:1139667240"        
#>  [85] "hb:300972942"          "hb:394225044"          "hb:5813067721"        
#>  [88] "hb:5813068729"         "hb:642664141"          "hb:673702721"         
#>  [91] "hb:457175171"          "hb:579916831"          "hb:673366098"         
#>  [94] "hb:5812981264"         "hb:5812981543"         "hb:457196444"         
#>  [97] "hb:5813020828"         "hb:423774471"          "hb:517854468"         
#> [100] "hb:264083994"          "hb:5813022274"        

# queries on classes respecting dataset idiosyncrasies
cf_ids(manc="/class:descend.+", flywire="/super_class:descend.+", expand = TRUE)
#> flywire [1303 ids]: 720575940624402173 720575940627055303 720575940631082808 720575940623037720 720575940629543409 720575940618666142 720575940638109668 720575940630697078 720575940610610841 720575940618870881 ...
#> manc [1328 ids]: 28478 31361 156265 19821 22072 24087 29182 30346 31742 33545 ...
# }
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [2 ids]: 2 1
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = 1:2, flywire=1:2), unique=FALSE)
#> flywire [3 ids]: 1 1 2
#> hemibrain [3 ids]: 2 1 2
# \donttest{
c(cf_ids(flywire = 1, hemibrain = 2), cf_ids(hemibrain = '/SMP55[0-9]', flywire=1:2))
#> flywire [2 ids]: 1 2
#> hemibrain [19 ids]: 2 452689494 579700707 328274638 392757094 605166418 298262663 361114148 669670654 613394707 ...
# }