Flexible function for fetching partner data across datasets

Flexible function for fetching partner data across datasets

Usage

cf_partners(
  ids,
  threshold = 1L,
  partners = c("inputs", "outputs"),
  bind.rows = TRUE
)

Arguments

ids

A list of ids named by the relevant datasets (see examples)

threshold

return only edges with at least this many matches. 0 is an option since neuprint sometimes returns 0 weight edges.

partners

Whether to return inputs or outputs

bind.rows

Whether to bind data.frames for each dataset together, keeping only the common columns (default TRUE for convenience but note that some columns will be dropped).

Value

A data.frame or a named list (when bind.rows=FALSE)

Examples

# \donttest{
da2ds=cf_partners(list(hemibrain='DA2_lPN'), threshold=5)
library(dplyr)
da2ds %>%
  count(type, wt=weight, sort=TRUE)
#>            type   n
#> 1       ORN_DA2 667
#> 2       DA2_lPN  89
#> 3         lLN10  85
#> 4         lLN2S  85
#> 5        v2LN30  72
#> 6  hb1796818119  68
#> 7       LHPV4b3  56
#> 8       lLN2T_a  48
#> 9       lLN2P_c  46
#> 10 hb1796817841  32
#> 11       l2LN18  23
#> 12      LHCENT3  14
#> 13 hb1797505019  14
#> 14          APL  12
#> 15 hb1827516355  12
#> 16      lLN2T_d  12
#> 17      lLN2T_e  11
#> 18      LHPD5c1  10
#> 19       lLN1_b  10
#> 20     LHPV12a1   6
#> 21          CSD   5
#> 22    LHPV2b5_a   5
#> 23  hb818983130   5
#> 24       lLN12A   5
#> 25       lLN1_c   5
# }
if (FALSE) {
cf_partners(list(flywire='DA2_lPN'))

DA2_lPN=cf_partners(list(flywire='DA2_lPN', malecns='DA2_lPN'))
}